chore: consalidate ssb calculation into one file
