+++ src/du_parameters/ssb.py
@@ -3,8 +3,13 @@
 
 """Synchronization Signal block calculations for different frequencies."""
 
+import logging
 from abc import ABC
-from typing import Union
+from typing import Optional, Union
+
+from src.du_parameters.afrcn import freq_to_arfcn
+
+logger = logging.getLogger(__name__)
 
 
 class BaseSSB(ABC):
@@ -172,3 +177,50 @@ def get_frequency_instance(frequency: Union[float, int]) -> BaseSSB:
             return frequency_cls(frequency)
 
     raise ValueError(f"Frequency {frequency} is out of supported range.")
+
+
+def get_absolute_frequency_ssb(center_freq: Union[int, float]) -> Optional[int]:
+    """Calculate absolute frequency for ssb using center frequency.
+
+    Args:
+        center_freq (float or int): Center frequency in MHz.
+
+    Returns:
+        arfcn (int): if successful, else None.
+    """
+    try:
+        try:
+            # Get frequency instance
+            frequency_instance = get_frequency_instance(center_freq)
+        except (ValueError, TypeError):
+            logger.error(f"Failed to create a frequency instance for center_freq={center_freq}")
+            return None
+
+        try:
+            # Calculate GSCN
+            gcsn = frequency_instance.freq_to_gscn()
+        except ValueError:
+            logger.error(f"Failed to calculate GSCN for center_freq={center_freq}")
+            return None
+
+        try:
+            # Convert GSCN to frequency
+            frequency_from_gcsn = frequency_instance.gscn_to_freq(gcsn)
+        except ValueError:
+            logger.error(f"Failed to calculate frequency using gcsn={gcsn}")
+            return None
+
+        try:
+            # Convert frequency to ARFCN
+            arfcn = freq_to_arfcn(frequency_from_gcsn)
+        except (ValueError, TypeError):
+            logger.error(f"Failed to calculate ARFCN for center_freq={center_freq}")
+            return None
+
+        return arfcn
+
+    except Exception as e:
+        logger.error(
+            f"Error in getting absolute frequency for ssb using center_freq={center_freq}: {e}"
+        )
+        return None
